Resolve conflicting push (#4)
* first stage of development reached (#1)

* adapting the keyscan, log and sdbdetect classes

* same as before, adapting old code

* reimplementing the keyscanning, not compiling yet

* reached state of old code, fully compining

* Update LICENSE

* reduced workload of keyscan, single file executable

* using array of keycodes

Reduced search space from 256 to 77 keys, drastically improving performance as the GetAsyncKeyState() function was mostly responsible for the processing time spent.

* Introducing parent module

In the future the system of obfuscation will be more advanced, creating multiple EXEs in the AppData folder with different names. Trying to acurately immitate a program install. Does that work?